Featue/check if time entries exist before delete project

+++ src/main/java/pl/betse/beontime/projectservice/service/ProjectService.java
@@ -2,13 +2,16 @@ package pl.betse.beontime.projectservice.service;
 
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 import pl.betse.beontime.projectservice.bo.ProjectBo;
 import pl.betse.beontime.projectservice.entity.ClientEntity;
 import pl.betse.beontime.projectservice.entity.ProjectEntity;
 import pl.betse.beontime.projectservice.exception.ClientNotFoundException;
 import pl.betse.beontime.projectservice.exception.ProjectAlreadyExistException;
 import pl.betse.beontime.projectservice.exception.ProjectNotFoundException;
+import pl.betse.beontime.projectservice.exception.TimeEntriesForProjectExists;
 import pl.betse.beontime.projectservice.mapper.ProjectMapper;
 import pl.betse.beontime.projectservice.repository.ClientRepository;
 import pl.betse.beontime.projectservice.repository.ProjectRepository;
@@ -24,6 +27,11 @@ public class ProjectService {
     private final ProjectRepository projectRepository;
     private final ClientRepository clientRepository;
     private final ProjectMapper projectMapper;
+
+    @Value("${api-project-exist}")
+    private String API_PROJECT_EXIST;
+
+    String tempPrefix = "http://localhost:8088";
 
     public ProjectService(ProjectRepository projectRepository, ClientRepository clientRepository, ProjectMapper projectMapper) {
         this.projectRepository = projectRepository;
@@ -75,9 +83,18 @@ public class ProjectService {
     }
 
     public void deleteProjectByGuid(String guid) {
-        ProjectEntity projectEntity = projectRepository.findByGuid(guid)
-                .orElseThrow(ProjectNotFoundException::new);
-        projectRepository.delete(projectEntity);
+        RestTemplate restTemplate = new RestTemplate();
+        String url = API_PROJECT_EXIST + guid;
+        boolean isTimeEntryForProjectExist = restTemplate.getForEntity(url, Boolean.class).getBody().booleanValue();
+        if (!isTimeEntryForProjectExist) {
+            ProjectEntity projectEntity = projectRepository.findByGuid(guid)
+                    .orElseThrow(ProjectNotFoundException::new);
+            projectRepository.delete(projectEntity);
+        } else {
+            log.error("Project with guid =" + guid + " cannot be deleted because time entries exist.");
+            throw new TimeEntriesForProjectExists();
+        }
+
     }
 
     private void updateProjectFields(ProjectBo projectBo, ProjectEntity projectEntity) {
